requirements.txt
Keras==2.1.6
jupyter==1.0.0
notebook==5.0.0
numpy==1.15.4
pandas==0.23.4
Django==2.0
django-rest-framework==0.1.0
